Feature/rcdy 286 Content, 비동기 리팩토링 (#25)
* chore: entity name and foreignKey name

* RCDY-286 chore: migration to movie detail

* RCDY-286 fix: movie entity and mappers

* RCDY-286 feat: search content feature with persisting

작품을 검색할 때 고유 id 를 부여함.

* RCDY-286 feat: movie search query from db

* RCDY-286 feat: catalog search from db

* RCDY-286 feat: BasicCategory migration to class

* RCDY-286 feat: registrar interfaces

* RCDY-286 refactor: basic category as class

* RCDY-286 refactor: divided genre resolver features

* RCDY-286 feat: entity, repository, mappers redesign

* RCDY-286 feat: entity registrar impls

* RCDY-286 feat: fetch movie credits

영화 배우 정보, 감독 정보를 가져온다.

* RCDY-286 feat: fetch, get movie details

api 에서 가져오는 경우와 db 에서 가져오는 경우를 나누었음.

* RCDY-286 feat: search, detail services

* RCDY-286 chore: minor change

* RCDY-286 test: test minor changes

+++ service-movie/api/src/main/java/com/recody/recodybackend/movie/features/getmoviecredit/GetMovieCredit.java
@@ -9,5 +9,5 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class GetMovieCredit {
     
-    private Long movieId;
+    private Integer movieId;
 }
